+++ src/Pages/Application/DeletePage.php
@@ -6,9 +6,7 @@ namespace Thinktomorrow\Chief\Pages\Application;
 
 use Illuminate\Support\Facades\DB;
 use Thinktomorrow\Chief\Modules\Module;
-use Thinktomorrow\Chief\Pages\Page;
 use Thinktomorrow\Chief\Audit\Audit;
-use Thinktomorrow\Chief\FlatReferences\FlatReferenceFactory;
 use Thinktomorrow\Chief\Urls\UrlRecord;
 use Thinktomorrow\Chief\States\PageState;
 
@@ -16,7 +14,6 @@ class DeletePage
 {
     public function handle($flatreference)
     {
-
         try {
             DB::beginTransaction();