+++ CHANGELOG.md
@@ -5,6 +5,7 @@ principles.
 ## unreleased
 - Added: A new page can be created based on an existing template. This will duplicate the pagebuilder content and structure.
 - Added: Pagefield used to select an internal page
+- Added: chief admin toast - a widget element on frontend for quick page edit and preview mode toggle
 - Added: dynamic attributes. [https://thinktomorrow.github.io/package-docs/src/chief/managers.html#dynamic-attributes](view docs)
 - Added: `Manager::editView()` and `Manager::createView` provide the option to customise the view filepath for the create and edit admin views.
 - Added: `Manager::editFields()` and `Manager::createFields` provide the fields for respectively the create and edit form views.
@@ -20,6 +21,7 @@ principles.
 - Changed: baseUrlSegment method is no longer a static method and requires to be an instance method.
 - Changed: Moved the Healthmonitor classes to the `src/System` directory. You'll need to update your chief.php monitor entries.
 - Fixed: Creating own modules for pages with custom tables now works as expected.
+- Removed: `ManagedModel::previewUrl()`. Use the `ManagedModel::url()` instead.
 - Removed: the `$manager` instance is no longer available inside a field view. You can always pass this to the view via the `Field::viewData()` method.
 - Removed: `Manager::fieldValue()` and `Manager::renderField()` methods. A field is now responsible for rendering its content, not the manager.
 - Removed: `RenderingFields` trait as the methods `Manager::fieldValue()` and `Manager::renderField()` are no longer being used.